LibPthread: Demonstrate memory leak and null deref
(1) Memory leak: Who deallocates the stack? The created thread cannot do it;
or it would need to munmap'ing the stack and call thread_exit without a stack.
The caller also cannot reasonably be asked to do so. glibc solves this problem
by placing the burden on the joiner thread; a solution I do not want to
implement.

(2) Null deref: The function at the base of the stack must not return,
because the return address is not meaningful.  Therefore, the function
given to pthread_create should not be used for the base of the stack.
On the other hand, this should not crash the kernel, so this is two bugs.

(3) User-visible modification of the attr structs: It appears that calling
pthread_create in a loop on the same non-null pthread_attr_t is an intended
use-case that Serenity should support. However, currently pthread_create may
allocate a stack and write the pointer to it into pthread_attr_t. This means
consecutive invocations will create threads which share a stack! This may
create problems when porting other programs.