LibPthread: Don't leak stack, control flow, or state
For a more detailed description of the problem, see the previous commit.

Note that this addresses the memory leak in case of a unsuccessul
syscall. The memory leak of the stack and thread-specific storage
need to be handled separately.

Furthermore, this changes the behavior in the case where `start_routine`
is an invalid pointer. Previously, pthread_create returned -EFAULT,
even though pthread_create specified no such return value. Instead,
this implementation allows the creation of the thread, which then promptly
SIGSEGV's itself by trying to execute the instruction at vaddr 0.

If it turns out that this was intended behavior after all,
pthread_create can just add a simple `if`.