AK: Add size checks to `AK::Statistics` to prevent bugs and errors
Previously many of the Methods in `AK::Statistics` would assume that
the size of `m_values` is non-zero. This could lead to zero divisions
and out of bounds reads.
For the Methods where an empty collection has a sensible result, this
is just returned. The other ones just assert the size is non-zero.